+++ packages/plugins/GoPlusSecurity/src/SNSAdaptor/components/RiskCard.tsx
@@ -2,7 +2,7 @@ import { memo, ReactNode } from 'react'
 import { Box, Stack, Typography } from '@mui/material'
 import { useI18N } from '../../locales/index.js'
 import { makeStyles } from '@masknet/theme'
-import type { GoPlusLabsSecurityMessage as SecurityMessage, SecurityAPI } from '@masknet/web3-providers'
+import type { SecurityMessage, SecurityAPI } from '@masknet/web3-providers'
 import { DefineMapping } from '../constants.js'
 
 const useStyles = makeStyles()((theme) => ({
