+++ packages/web3-providers/src/index.ts
@@ -51,12 +51,7 @@ import { MulticallAPI } from './Multicall/index.js'
 import { ENS_API } from './ENS/index.js'
 import { SpaceID_API } from './SpaceID/index.js'
 
-// #region TODO: to be removed
 export * from './types/index.js'
-export * from './OpenSea/utils.js'
-export type { SecurityMessage as GoPlusLabsSecurityMessage } from './GoPlusLabs/rules.js'
-export { CollectionType } from './RSS3/constants.js'
-// #endregion
 
 export const OpenSea = new OpenSeaAPI()
 export const LooksRare = new LooksRareAPI()
